Copied from example 14.11
I do not fully understand Jquery or even JavaScript. I plan on asking my teacher about it soon. However this doesn't change the fact that I have a deadline coming up. So figure I'll adapt this so it'll work with my images, and then come back later. My Professor has said in the past that we can use these if we have too much trouble.